requirements.txt
# Core Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-dotenv>=1.0.0

# Data Analysis and Visualization
matplotlib>=3.7.0
pandas>=2.0.0
python-dateutil>=2.8.0

# Template Engine
Jinja2>=3.1.0

# Secure Gmail API Integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

# Email and SMTP
# Note: smtplib and email.mime are built-in Python modules

# Database
# Note: sqlite3 is built-in Python module

# Configuration
# Note: configparser is built-in Python module